remove inflation
