+++ Content.Goobstation.Server/ServerCurrency/Commands/ServerCurrencyCommands.cs
@@ -13,8 +13,9 @@ namespace Content.Goobstation.Server.ServerCurrency.Commands;
 [AnyCommand]
 public sealed partial class BalanceServerCurrencyCommand : IConsoleCommand
 {
-    [Dependency] private ICommonCurrencyManager _currencyMan = default!;
-    [Dependency] private IChatManager _chatManager = default!;
+    [Dependency] private ICommonCurrencyManager _currency = default!;
+    [Dependency] private IChatManager _chat = default!;
+
     public string Command => Loc.GetString("server-currency-balance-command");
     public string Description => Loc.GetString("server-currency-balance-command-description");
     public string Help => Loc.GetString("server-currency-balance-command-help");
@@ -33,9 +34,9 @@ public sealed partial class BalanceServerCurrencyCommand : IConsoleCommand
         }
 
         var balance = Loc.GetString("server-currency-balance-command-return",
-            ("balance", _currencyMan.Stringify(_currencyMan.GetBalance(shell.Player.UserId))));
+            ("balance", _currency.Stringify(_currency.GetBalance(shell.Player.UserId))));
 
-        _chatManager.ChatMessageToOne(ChatChannel.Local, balance, balance, EntityUid.Invalid, false, shell.Player.Channel);
+        _chat.ChatMessageToOne(ChatChannel.Local, balance, balance, EntityUid.Invalid, false, shell.Player.Channel);
         shell.WriteLine(balance);
     }
 }
@@ -43,8 +44,9 @@ public sealed partial class BalanceServerCurrencyCommand : IConsoleCommand
 [AnyCommand]
 public sealed partial class GiftServerCurrencyCommand : IConsoleCommand
 {
-    [Dependency] private ICommonCurrencyManager _currencyMan = default!;
-    [Dependency] private IChatManager _chatManager = default!;
+    [Dependency] private ICommonCurrencyManager _currency = default!;
+    [Dependency] private IChatManager _chat = default!;
+    [Dependency] private IPlayerManager _player = default!;
 
     public string Command => Loc.GetString("server-currency-gift-command");
     public string Description => Loc.GetString("server-currency-gift-command-description");
@@ -58,17 +60,18 @@ public sealed partial class GiftServerCurrencyCommand : IConsoleCommand
             return;
         }
 
-        if(shell.Player is not { } player){
+        if (shell.Player is not { } player)
+        {
             shell.WriteError(Loc.GetString("shell-cannot-run-command-from-server"));
             return;
         }
 
-        var plyMgr = IoCManager.Resolve<IPlayerManager>();
-        if (!plyMgr.TryGetUserId(args[0], out var targetPlayer))
+        if (!_player.TryGetUserId(args[0], out var targetPlayer))
         {
             shell.WriteError(Loc.GetString("server-currency-command-error-1"));
             return;
-        } else if (targetPlayer == shell.Player.UserId)
+        }
+        else if (targetPlayer == shell.Player.UserId)
         {
             shell.WriteError(Loc.GetString("server-currency-gift-command-error-1"));
             return;
@@ -85,19 +88,20 @@ public sealed partial class GiftServerCurrencyCommand : IConsoleCommand
         if (amount == 0)
             amount = 1; // Trolled
 
-        if (!_currencyMan.CanAfford(shell.Player.UserId, amount, out int balance)){
+        if (!_currency.CanAfford(shell.Player.UserId, amount, out int balance))
+        {
             shell.WriteError(Loc.GetString("server-currency-gift-command-error-2", ("balance", balance)));
             return;
         }
 
-        _currencyMan.TransferCurrency(shell.Player.UserId, targetPlayer, amount);
+        _currency.TransferCurrency(shell.Player.UserId, targetPlayer, amount);
 
-        var giver = Loc.GetString("server-currency-gift-command-giver", ("player", args[0]), ("amount", _currencyMan.Stringify(amount)));
-        var reciever = Loc.GetString("server-currency-gift-command-reciever", ("player", shell.Player.Name), ("amount", _currencyMan.Stringify(amount)));
+        var giver = Loc.GetString("server-currency-gift-command-giver", ("player", args[0]), ("amount", _currency.Stringify(amount)));
+        var reciever = Loc.GetString("server-currency-gift-command-reciever", ("player", shell.Player.Name), ("amount", _currency.Stringify(amount)));
 
-        if(plyMgr.TryGetSessionById(targetPlayer, out var targetPlayerSession))
-            _chatManager.ChatMessageToOne(ChatChannel.Local, reciever, reciever, EntityUid.Invalid, false, targetPlayerSession.Channel);
-        _chatManager.ChatMessageToOne(ChatChannel.Local, giver, giver, EntityUid.Invalid, false, shell.Player.Channel);
+        if (_player.TryGetSessionById(targetPlayer, out var targetPlayerSession))
+            _chat.ChatMessageToOne(ChatChannel.Local, reciever, reciever, EntityUid.Invalid, false, targetPlayerSession.Channel);
+        _chat.ChatMessageToOne(ChatChannel.Local, giver, giver, EntityUid.Invalid, false, shell.Player.Channel);
 
         shell.WriteLine(giver);
     }
@@ -116,7 +120,8 @@ public sealed partial class GiftServerCurrencyCommand : IConsoleCommand
 [AdminCommand(AdminFlags.Host)]
 public sealed partial class AddServerCurrencyCommand : IConsoleCommand
 {
-    [Dependency] private ICommonCurrencyManager _currencyMan = default!;
+    [Dependency] private ICommonCurrencyManager _currency = default!;
+    [Dependency] private IPlayerManager _player = default!;
 
     public string Command => Loc.GetString("server-currency-add-command");
     public string Description => Loc.GetString("server-currency-add-command-description");
@@ -130,8 +135,7 @@ public sealed partial class AddServerCurrencyCommand : IConsoleCommand
             return;
         }
 
-        var plyMgr = IoCManager.Resolve<IPlayerManager>();
-        if (!plyMgr.TryGetUserId(args[0], out var targetPlayer))
+        if (!_player.TryGetUserId(args[0], out var targetPlayer))
         {
             shell.WriteError(Loc.GetString("server-currency-command-error-1"));
             return;
@@ -143,7 +147,7 @@ public sealed partial class AddServerCurrencyCommand : IConsoleCommand
             return;
         }
 
-        var newCurrency = _currencyMan.Stringify(_currencyMan.AddCurrency(targetPlayer, currency));
+        var newCurrency = _currency.Stringify(_currency.AddCurrency(targetPlayer, currency));
         shell.WriteLine(Loc.GetString("server-currency-command-return", ("player", args[0]), ("balance", newCurrency)));
     }
 
@@ -161,7 +165,8 @@ public sealed partial class AddServerCurrencyCommand : IConsoleCommand
 [AdminCommand(AdminFlags.Host)]
 public sealed partial class RemoveServerCurrencyCommand : IConsoleCommand
 {
-    [Dependency] private ICommonCurrencyManager _currencyMan = default!;
+    [Dependency] private ICommonCurrencyManager _currency = default!;
+    [Dependency] private IPlayerManager _player = default!;
 
     public string Command => Loc.GetString("server-currency-remove-command");
     public string Description => Loc.GetString("server-currency-remove-command-description");
@@ -175,8 +180,7 @@ public sealed partial class RemoveServerCurrencyCommand : IConsoleCommand
             return;
         }
 
-        var plyMgr = IoCManager.Resolve<IPlayerManager>();
-        if (!plyMgr.TryGetUserId(args[0], out var targetPlayer))
+        if (!_player.TryGetUserId(args[0], out var targetPlayer))
         {
             shell.WriteError(Loc.GetString("server-currency-command-error-1"));
             return;
@@ -188,7 +192,7 @@ public sealed partial class RemoveServerCurrencyCommand : IConsoleCommand
             return;
         }
 
-        var newCurrency = _currencyMan.Stringify(_currencyMan.RemoveCurrency(targetPlayer, currency));
+        var newCurrency = _currency.Stringify(_currency.RemoveCurrency(targetPlayer, currency));
         shell.WriteLine(Loc.GetString("server-currency-command-return", ("player", args[0]), ("balance", newCurrency)));
     }
 
@@ -206,7 +210,8 @@ public sealed partial class RemoveServerCurrencyCommand : IConsoleCommand
 [AdminCommand(AdminFlags.Host)]
 public sealed partial class SetServerCurrencyCommand : IConsoleCommand
 {
-    [Dependency] private ICommonCurrencyManager _currencyMan = default!;
+    [Dependency] private ICommonCurrencyManager _currency = default!;
+    [Dependency] private IPlayerManager _player = default!;
 
     public string Command => Loc.GetString("server-currency-set-command");
     public string Description => Loc.GetString("server-currency-set-command-description");
@@ -220,8 +225,7 @@ public sealed partial class SetServerCurrencyCommand : IConsoleCommand
             return;
         }
 
-        var plyMgr = IoCManager.Resolve<IPlayerManager>();
-        if (!plyMgr.TryGetUserId(args[0], out var targetPlayer))
+        if (!_player.TryGetUserId(args[0], out var targetPlayer))
         {
             shell.WriteError(Loc.GetString("server-currency-command-error-1"));
             return;
@@ -233,8 +237,8 @@ public sealed partial class SetServerCurrencyCommand : IConsoleCommand
             return;
         }
 
-        _currencyMan.SetBalance(targetPlayer, currency);
-        var newCurrency = _currencyMan.Stringify(currency);
+        _currency.SetBalance(targetPlayer, currency);
+        var newCurrency = _currency.Stringify(currency);
         shell.WriteLine(Loc.GetString("server-currency-command-return", ("player", args[0]), ("balance", newCurrency)));
     }
 
@@ -252,7 +256,8 @@ public sealed partial class SetServerCurrencyCommand : IConsoleCommand
 [AdminCommand(AdminFlags.Host)]
 public sealed partial class GetServerCurrencyCommand : IConsoleCommand
 {
-    [Dependency] private ICommonCurrencyManager _currencyMan = default!;
+    [Dependency] private ICommonCurrencyManager _currency = default!;
+    [Dependency] private IPlayerManager _player = default!;
 
     public string Command => Loc.GetString("server-currency-get-command");
     public string Description => Loc.GetString("server-currency-get-command-description");
@@ -266,14 +271,13 @@ public sealed partial class GetServerCurrencyCommand : IConsoleCommand
             return;
         }
 
-        var plyMgr = IoCManager.Resolve<IPlayerManager>();
-        if (!plyMgr.TryGetUserId(args[0], out var targetPlayer))
+        if (!_player.TryGetUserId(args[0], out var targetPlayer))
         {
             shell.WriteError(Loc.GetString("server-currency-command-error-1"));
             return;
         }
 
-        var currency = _currencyMan.Stringify(_currencyMan.GetBalance(targetPlayer));
+        var currency = _currency.Stringify(_currency.GetBalance(targetPlayer));
         shell.WriteLine(Loc.GetString("server-currency-command-return", ("player", args[0]), ("balance", currency)));
     }
 
@@ -282,6 +286,37 @@ public sealed partial class GetServerCurrencyCommand : IConsoleCommand
         return args.Length switch
         {
             1 => CompletionResult.FromHintOptions(CompletionHelper.SessionNames(), Loc.GetString("server-currency-command-completion-1")),
+            _ => CompletionResult.Empty
+        };
+    }
+}
+
+[AdminCommand(AdminFlags.Host)]
+public sealed partial class WipeServerCurrencyCommand : IConsoleCommand
+{
+    [Dependency] private ICommonCurrencyManager _currency = default!;
+
+    public string Command => "wipecurrency";
+    public string Description => "Wipe the entire server's currency database...";
+    public string Help => "wipecurrency [codephrase]";
+
+    public async void Execute(IConsoleShell shell, string argStr, string[] args)
+    {
+        if (args.Length != 1 || args[0] != "La-li-lu-le-lo")
+        {
+            shell.WriteError("EVA, say the password.");
+            return;
+        }
+
+        await _currency.Wipe();
+        shell.WriteLine("You're face, to face, with the man who sold the world");
+    }
+
+    public CompletionResult GetCompletion(IConsoleShell shell, string[] args)
+    {
+        return args.Length switch
+        {
+            1 => CompletionResult.FromHint("Who are the patriots?"),
             _ => CompletionResult.Empty
         };
     }
